requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
flask==3.0.0